nixos/virtualisation.oci-containers: add systemd 'after' and 'requires' options
Currently, enabling `autostart` on containers declared in `virtualisation.oci-containers.containers` does not wait for network to be ready. It is possible that such containers cannot pull images from a container registry, which causes the services to fail. This change makes it possible to declare systemd units such as `network.target` to postpone the unit startup.